Possible mod to upper active user display... MIGHT fix?  Can't test locally.
XML created at end modified:  nothing to do in the case of "html_content" which is likely the case most of the time (during the iterations through the list).  "success_message" or "default" may have sometimes been cases, just not when the data was collected through use of the list.  My GUESS is this will repair the display.